requirements.txt
Django ~= 4.2.3
mysqlclient ~= 2.2.0
django-crispy-forms ~=2.0
crispy-bootstrap5 ~= 0.7
pycryptodome ~= 3.18.0
django-environ ~= 0.10.0
